+++ .eslintrc.js
@@ -5,7 +5,7 @@ module.exports = {
     "jest/globals": true,
   },
   parserOptions: {
-    ecmaVersion: 2018,
+    ecmaVersion: 2019,
   },
   extends: ["eslint:recommended", "prettier"],
   plugins: ["jest"],