+++ src/__tests__/converters.js
@@ -1,9 +1,11 @@
 const data = require(`./data.json`)
 
-const convertGoogleDocumentToJson = require("../utils/convert-google-document-to-json")
-const convertJsonToMarkdown = require("../utils/convert-json-to-markdown")
+const {
+  convertGoogleDocumentToJson,
+  convertJsonToMarkdown,
+} = require("../utils/converters")
 
-test("converts document", () => {
+test("Converts document", () => {
   const file = {id: `hi`, metadatIsCool: `yeah`}
   const content = convertGoogleDocumentToJson(data)
   expect(convertJsonToMarkdown({file, content})).toMatchSnapshot()