+++ src/utils/convert-json-to-markdown.js
@@ -1,9 +1,0 @@
-const json2md = require("json2md")
-const YAML = require("yamljs")
-
-module.exports = ({content, file}) => {
-  return `---
-${YAML.stringify(file)}---
-
-${json2md(content)}`
-}